Updated to match new design
First a random quote will be shown with the login icon being shown in the bottom right. When clicking on the screen the login UI will enlarge to the normal size. During this transition the quote will fade and the background will blur. Now the sign up text is within the login hexagon which still leads to the sign up menu. A new button was made to allow access back to the login menu.